requirements.txt
omegaconf>=2.3.0
hydra-core>=1.3.2
hydra-callbacks>=0.6.1
scikit-learn>=1.5.1
lightning>=2.4.0
tensorboard>=2.17.1
imops==0.9.0
monai==1.3.2
medimm==0.0.3